Update i-stopped-using-visual-studio-code.md
I couldnt comment on your post, so I figured go the github route.
It'd be good to elaboraye a little more on this vs that tool. almost feels like a gut feel, the way the article is written, which is also a good way to go about it. Doesnt help me much though.